feat: checkpoint current workspace changes
