+++ .env.example
@@ -7,6 +7,8 @@ NEXTAI_API_KEY=
 NEXTAI_WEB_DIR=web
 NEXTAI_ACTIVE_PROVIDER=demo
 NEXTAI_ACTIVE_MODEL=demo-chat
+NEXTAI_ENABLE_PROMPT_TEMPLATES=false
+NEXTAI_ENABLE_PROMPT_CONTEXT_INTROSPECT=false
 
 # Optional tools
 NEXTAI_ENABLE_BROWSER_TOOL=false
