+++ apps/web/src/index.html
@@ -4,6 +4,15 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>NextAI 控制台</title>
+    <link
+      rel="icon"
+      type="image/svg+xml"
+      href="data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 64 64'%3E%3Crect width='64' height='64' rx='14' fill='%231f2a37'/%3E%3Cpath d='M18 46V18h8l20 20V18h8v28h-8L26 26v20z' fill='%23f5f6f8'/%3E%3C/svg%3E"
+    />
+    <link
+      rel="apple-touch-icon"
+      href="data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 64 64'%3E%3Crect width='64' height='64' rx='14' fill='%231f2a37'/%3E%3Cpath d='M18 46V18h8l20 20V18h8v28h-8L26 26v20z' fill='%23f5f6f8'/%3E%3C/svg%3E"
+    />
     <link rel="stylesheet" href="./styles.css" />
   </head>
   <body>
@@ -519,7 +528,7 @@
                     </div>
                   </div>
                   <div class="composer-meta-row" aria-live="polite">
-                    <span id="composer-token-estimate" class="composer-token-estimate">Tokens: 0</span>
+                    <span id="composer-token-estimate" class="composer-token-estimate">0/128.0k</span>
                   </div>
                   <div class="composer-status-row">
                     <button type="button" class="composer-status-btn">
@@ -725,5 +734,3 @@
     <script type="module" src="./main.js"></script>
   </body>
 </html>
-
-
