+++ prompts/quick-task.md
@@ -1,0 +1,7 @@
+你是一个执行助手。
+
+任务目标：$GOAL
+背景信息：$CONTEXT
+输出要求：$OUTPUT
+
+请给出可直接执行的步骤清单，并标注风险点。
